+++ api/konnect/v1alpha1/zz_generated_portalteam_funcs.go
@@ -104,3 +104,14 @@ func (obj *PortalTeam) GetStatusConditionReasonParentRefInvalid() string {
 func (obj *PortalTeam) GetStatusConditionReasonParentRefNotProgrammed() string {
 	return PortalRefReasonNotProgrammed
 }
+
+// GetAncestorIDs returns the Konnect IDs of the ancestor entities keyed by their Kind.
+func (obj *PortalTeam) GetAncestorIDs() map[string]string {
+	m := make(map[string]string, 1)
+	if obj.Status.PortalID != nil {
+		m["Portal"] = obj.Status.PortalID.ID
+	} else {
+		m["Portal"] = ""
+	}
+	return m
+}