+++ controller/konnect/reconciler_serviceref_test.go
@@ -50,7 +50,7 @@ var testKongServiceOK = &configurationv1alpha1.KongService{
 		},
 	},
 	Status: configurationv1alpha1.KongServiceStatus{
-		Konnect: &konnectv1alpha2.KonnectEntityStatusWithControlPlaneRef{
+		Konnect: &konnectv1alpha2.KonnectEntityStatusWithControlPlaneAndCertificateAndCACertificatesRefs{
 			KonnectEntityStatus: konnectv1alpha2.KonnectEntityStatus{
 				ID: "12345",
 			},
@@ -79,7 +79,7 @@ var testKongServiceWithCPRefUnprogrammed = &configurationv1alpha1.KongService{
 		},
 	},
 	Status: configurationv1alpha1.KongServiceStatus{
-		Konnect: &konnectv1alpha2.KonnectEntityStatusWithControlPlaneRef{
+		Konnect: &konnectv1alpha2.KonnectEntityStatusWithControlPlaneAndCertificateAndCACertificatesRefs{
 			KonnectEntityStatus: konnectv1alpha2.KonnectEntityStatus{
 				ID: "12345",
 			},
@@ -164,7 +164,7 @@ var testKongServiceCrossNs = &configurationv1alpha1.KongService{
 		},
 	},
 	Status: configurationv1alpha1.KongServiceStatus{
-		Konnect: &konnectv1alpha2.KonnectEntityStatusWithControlPlaneRef{
+		Konnect: &konnectv1alpha2.KonnectEntityStatusWithControlPlaneAndCertificateAndCACertificatesRefs{
 			KonnectEntityStatus: konnectv1alpha2.KonnectEntityStatus{
 				ID: "svc-cross-ns-id",
 			},